Add files via upload
Dataset

code_using_fandom.py extracts specific traits and personality text from Fandom, and turns them into a dataset that we can use. [Example: DF1.csv]
Summarizer.py extracts the description of the character from Fandom and gives a summary of the background and traits of the characters. [Example: DF2.csv]
merged.py merges the output of the data frames, cleans a few columns that we don't need and gives the final dataframe. [Example: DF3.csv]